Fix logs for video-recording (#184)
The log files "video-rec-stderr.log" and "video-rec-stdout.log" seem to contain "^M" characters (carriage-return).
The 'tail' command below doesn't take into account this character. It returns the complete output of the log files. Therefore it needs to be replaced into a proper newline.
This problem is not visible if executed in bash because each line will overwrite the previous line.